PLANNER-1224: Remove Save Button and Update Rotation REST Service API (#208)
There was a bug that can be reproduced by switching between and modifying the rotation and avaliability screens, causing an app freeze due to an "non-existing" node. This was probably caused by a race condition between updating a screen and changing the screen; if you change the screen first, the old screen elements are removed, which in turn cause "existing" elements to not have an element.

The solution is simple: for batch updates, remove all data and re-add data from the batch. The old way tries to be optimal by deleting removed elements and modifying current ones, to minimize changes to the DOM. The new way creates new lanes and new elements each time.

I have not observed a performance lost due to this change, and it appears the bug is fixed.