moment: bring in all locales
Locales influence the first day of the week, amongst other things.
While the US, Canada and Japan start the week on Sunday, many other
places start it on Monday.

To avoid ambiguity, bring all locales into moment. This results in
an inconsistency between moment and date-fns (and react-datepicker),
which we can reasonably bear with for now