Fix updating objects for Shifts and Employees, update Color Picker
The Shift issue was brought to our attention by
GitHub user @cgrgrbz PR:
https://github.com/kiegroup/optaweb-employee-rostering/pull/460
That PR had an accidental second commit that included custom
modifications. Once it was brought to their attention, they
closed the PR and deleted the branch, and haven't created
a PR containing only the fix. Since I needed to fix other
issues of the same nature, I decided to also include their
fix in this commit.