Unbind the dependency:analyze-only execution
The dependency:analyze-only execution is used
to discover usage of transitive dependencies that
are not declared, which can cause build issues
if the dependency that brought it in was removed
or replaced that dependency with something else.
Spring Start applications are different; their
transitive dependencies are meant to be relied
on, which simplify dependency management. Since
we aren't using any dependencies outside the
Spring ecosystem, the dependency plugin
brings an extra maintence burden with no
benefit.