PLANNER-2200: Update OptaPlanner version to 8.0.0-SNAPSHOT (#508)
* PLANNER-2200: Update OptaPlanner version to 8.0.0-SNAPSHOT

- Replace WannabeSolverManager with ScoreManager + SolverManager
- Replace OptaWeb's SolverStatus with OptaPlanner's SolverStatus
- Genericify Indictment usage
- Add scheduleUpdateOfRoster method to update shifts in a
  separate thread to not block Solver

* Change parent pom to jboss-parent

* Add missing property version; update java version

* PLANNER-2195: Cache maven and npm dependencies; update java version

* Merge Jiri suggestions

Drools solver seem to be broken; getting a NPE within drools

* PLANNER-2201: Change parent to optaplanner-build-parent

* Unbind the dependency:analyze-only execution

The dependency:analyze-only execution is used
to discover usage of transitive dependencies that
are not declared, which can cause build issues
if the dependency that brought it in was removed
or replaced that dependency with something else.
Spring Start applications are different; their
transitive dependencies are meant to be relied
on, which simplify dependency management. Since
we aren't using any dependencies outside the
Spring ecosystem, the dependency plugin
brings an extra maintence burden with no
benefit.

* Fix AsciiDoc plugin configuration

Due to changing the build parent, the
AsciiDoc Maven Plugin version updated
from 2.0.0-RC.1 to 2.0.0, which
introduced some breaking changes.
Notably, the images directory and
source highlighter must be configured
as an attribute. For more details
on what changed, see
https://github.com/asciidoctor/asciidoctor-maven-plugin/blob/master/docs/v2-migration-guide.adoc

* PLANNER-2189: Do not start a new refresh timeout in refresh

Also includes sonarcloud bug/code smell fixes