Configure docker to run as non-root user
I need to run docker with non-root so to be able to delete files generated inside the container.

Though for podman it is opposite - if you run it as a non-root it is unable to write to the binded volume

I used here 1000:1000 as user and group id since it is a default non-root. So it exist both in container and on host so to not get noisy notifications like:

"groups: cannot find name for group ID"
"I have no name!"
  - your container, complaining