Make "employeeNotRotationEmployee" not trigger if shift.getRotationEmployee() is null
There was a difference between scoreDrl and constraint stream
implementations that was not tested in tests. Namely, when
the rotation employee is null and the employee is not null.
The CS impl triggers a penalty, whereas the DRL does not
trigger a penalty. The DRL behavior is the correct one.

As a result of switching to CS by default, this triggers
a NPE on frontend since it creates a employeeNotRotationEmployee
violation with a null rotation employee (which the frontend
assume must not be null).

This adds a test and make CS behave the same as DRL