Fix issue #776: Link contract violation constraint matches from employee to shift
Due to how justification lists in Constraint Stream works, Shift
indictments never included contract violation constraint matches
(since the shift is not in their justification list because the shifts
were removed in a groupBy operation).

Because the justification list still have the employee, we can use
the indictment of the employee to gets its contract violation constraint
matches, and add them to the shift's constraint matches.